Reworked version checking so that it no longer requires the checkVersions option. Now, the version will only be fetched from npm if "latest" is specified for the version. Then all versions will be tested for being semantically correct according to semvar. Finally, if the downloader receives a 404 error for the requested version we can give a helpful error explaining that they must specify a version available on the nodewebkit build server.
This has the advantage that in the case that the proper binaries are cached and version is explicitly specified, no request will be made because it is not necessary. Also, we have confined the version dependency down to simply the versions available at http://dl.node-webkit.org/. People can use RC builds or major/minor versions and no filtering of the master build list is necessary.