Add `mapChunked` function
`mapChunked` allows to map over a sequence by chunking the sequence into
chunks of maximum size `chunksize`. All values in a chunk are executed
in parallel and once the chunk as a whole is completed, the next chunk
will become subject to processing.

The function `f` is called on each element in a chunk and must return
a Promise.

The values in array `xs` can be of any type.