Do not expand `files`, add `fileglob`
The first argument to `utils.getFileList` is now a list of already
expanded files, the second argument is a list of fileglob patterns.

Boths lists are concatenated after expansion and made unique.

This allows a user to specify a set of known files or have node webkit
builder expand it on the fly. This is particularily interesting for
grunt, which does the file expansion already.

This changes the `NwBuilder` interface with a backwards incompatible
manner: The constructor now takes both a `files` and a `fileglob`
array. However, what is now called `fileglob` was previously called
`files` and will *NOT* be expanded further.